Move search list into molecules
