+++ packages/components/src/index.ts
@@ -27,8 +27,8 @@ export { default as ChannelMessageHeader } from './local/molecules/channel-messa
 export { default as ChannelMessageDetail } from './local/molecules/channel-message-detail'
 export { default as ChannelMessage } from './local/molecules/channel-message'
 export { default as ButtonList } from './local/molecules/button-list'
+export { default as SearchList } from './local/molecules/search-list'
 export { default as ChannelMessagesList } from './local/organisms/channel-messages-list'
 export { default as AdminMessagesList } from './local/organisms/admin-messages-list'
-export { default as SearchList } from './local/organisms/search-list'
 export { default as WargameList } from './local/organisms/wargame-list'
 export { default as AdminLogin } from './local/organisms/admin-login'
