+++ packages/components/src/local/molecules/search-list/styles.module.scss
@@ -1,4 +1,4 @@
-@import "~@serge/themes/dependencies/variables";
+@import "../../../../../../node_modules/@serge/themes/dependencies/variables";
 
 .searchlist {
   @include typo-primary;
